+++ server/src/main/java/com/cn/server/module/scene/entity/SceneId.java
@@ -5,21 +5,21 @@ public class SceneId {
 	/**
 	 * 启始之地
 	 */
-	public static final int QI_SHI_ZHI_DI=10012;
+	public static final short QI_SHI_ZHI_DI=10012;
 	
 	/**
 	 * 村子
 	 */
-	public static final int CUN_ZI=10013;
+	public static final short CUN_ZI=10013;
 	
 	/**
 	 * 森林
 	 */
-	public static final int SEN_LIN=10014;
+	public static final short SEN_LIN=10014;
 	
 	/**
 	 * 城堡
 	 */
-	public static final int CHENG_BAO=10015;
+	public static final short CHENG_BAO=10015;
 
 }
